auth root file build error fix
