+++ client/src/types/nextAuth.ts
@@ -1,5 +1,24 @@
 export interface KeycloakToken {
-    realm_access?: {
-      roles: string[];
-    };
+  exp: number;
+  iat: number;
+  auth_time: number;
+  jti: string;
+  iss: string;
+  aud: string[];
+  sub: string;
+  typ: string;
+  azp: string;
+  session_state: string;
+  acr: string;
+  realm_access: {
+    roles: string[];
+  };
+  scope: string;
+  sid: string;
+  email_verified: boolean;
+  name: string;
+  preferred_username: string;
+  given_name: string;
+  family_name: string;
+  email: string;
 }
